HACK: suppress clang-tidy warnings
These are quick hacks to silence clang-tidy warnings. The purpose of
this patch is just to highlight the warnings on the mailing list, and
also, if you want to try to fix some of the warnings, it may be nice to
have all the rest suppressed.

Afaics, the thread->moveObject(this) warning is a false positive.
Probably some of the others are too.

Note that I used assert() there in a few places, just after ASSERT().
The reason being, ASSERT() just prints an error, while assert() aborts,
so clang-tidy takes the assert() seriously.

Not for merging.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@iki.fi>